Migrate Window and Tabs.
