+++ src/components/ADempiere/DataTable/index.vue
@@ -1079,8 +1079,8 @@ export default {
     },
     getFieldDefinition(fieldDefinition, row) {
       var styleSheet = ''
-      if (fieldDefinition && (fieldDefinition.id !== null || fieldDefinition.conditions.length)) {
-        fieldDefinition.conditions.forEach(condition => {
+      if (fieldDefinition && (fieldDefinition.id !== null || fieldDefinition.conditionsList.length)) {
+        fieldDefinition.conditionsList.forEach(condition => {
           var columns = evaluator.parseDepends(condition.condition)
           var conditionLogic = condition.condition
           columns.forEach(column => {
